Editable CSS fixes
Makes min-height calculations work properly. Prior to this commit, the
min-height would get stuck on a certain value when doing SSR, and maybe
also without SSR in some scenarios, because the computed min-height
value was being used to set the new one.

Instead of using inline styles to set min-height and other properties
like position, outline, etc. a stylesheet is created and it is briefly
removed before updating min-height.